Fix Load<T> not calling event if asset doesn't exist
Not sure if this works or not since I just kinda threw it together in the GitHub text editor, but this should make Load<T> not throw an error if the asset doesn't exist in the game's Content directory, but instead pass default(T) to the event handlers and return the default value. This won't make it throw an error if nothing handles the asset though, but I'm not sure there's an easy fix for that.